Implement more OOP to improve Code Quality
Moving command operations logic from Storage class to different Command classes.

Creating an abstract command class and extracting more classes that inherits from it to handle the command from user input.
* AddCommand contains add operation for Todo, Deadline and Event task.
* ClearCommand contains clear all tasks operation.
* DeleteCommand contains delete one task operation.
* DoneCommand contains the operation of marking task as done.
* ExitCommand contains operation of showing exit message.
* FindCommand contains find matching tasks operation.
* ListCommand contains list all tasks operation.

Creating a Parser class to return the respective Command object.
