+++ src/main/java/duke/controller/DukeController.java
@@ -2,11 +2,13 @@ package duke.controller;
 
 import java.io.IOException;
 
+import duke.command.Command;
 import duke.controller.dialog.DukeDialogController;
 import duke.controller.dialog.UserDialogController;
 import duke.listener.Message;
 import duke.storage.Storage;
-import duke.task.Operation;
+import duke.task.TaskList;
+import duke.util.Parser;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.scene.control.ScrollPane;
@@ -33,12 +35,14 @@ public class DukeController extends AnchorPane implements Message {
     @FXML
     private ImageView ivSend;
 
+    private final TaskList taskList;
     private final Storage storage;
 
     /**
      * Constructs a DukeController object.
      */
     public DukeController() {
+        taskList = new TaskList();
         storage = new Storage(this);
     }
 
@@ -62,7 +66,8 @@ public class DukeController extends AnchorPane implements Message {
                                     content, preDukeDialogController.getHeight());
                             dialogContainer.getChildren().add(dukeDialogController);
                         } catch (IOException e) {
-                            System.out.println("Duke dialog fxml file not found: " + e.getMessage());
+                            System.out.println("Duke dialog fxml file not found: "
+                                    + e.getMessage());
                         } finally {
                             dialogContainer.getChildren().remove(preDukeDialogController);
                         }
@@ -94,7 +99,8 @@ public class DukeController extends AnchorPane implements Message {
                                     content, preUserDialogController.getHeight());
                             dialogContainer.getChildren().add(userDialogController);
                         } catch (IOException e) {
-                            System.out.println("User dialog fxml file not found: " + e.getMessage());
+                            System.out.println("User dialog fxml file not found: "
+                                    + e.getMessage());
                         } finally {
                             dialogContainer.getChildren().remove(preUserDialogController);
                         }
@@ -125,34 +131,24 @@ public class DukeController extends AnchorPane implements Message {
     }
 
     private void init() {
-        storage.loadTasks();
+        storage.loadTasks(taskList);
         addDukeDialog(GREET);
         tfInput.setDisable(false);
     }
 
     private void handleInput() {
         String input = tfInput.getText();
-        addUserDialog(input);
-        Operation operation = storage.getOperation(input);
-        if (operation == Operation.BYE) {
-            exit();
-        } else if (operation == Operation.LIST) {
-            storage.listTasks();
-        } else if (operation == Operation.DONE) {
-            storage.completeTask(input);
-        } else if (operation == Operation.DELETE) {
-            storage.deleteTask(input);
-        } else if (operation == Operation.CLEAR) {
-            storage.clearTasks();
-        } else if (operation == Operation.FIND) {
-            storage.findTasks(input);
-        } else if (operation == Operation.TODO
-            || operation == Operation.DEADLINE
-            || operation == Operation.EVENT) {
-            storage.addTask(input);
-        }
-        storage.saveTasksToFile();
         tfInput.clear();
+        addUserDialog(input);
+        Command c = Parser.parse(input, this);
+        if (c == null) {
+            return;
+        }
+        c.execute(taskList);
+        if (c.isExit()) {
+            exit();
+        }
+        storage.saveTasksToFile(taskList);
     }
 
     private void exit() {
