+++ src/main/java/duke/task/TaskList.java
@@ -11,8 +11,8 @@ import duke.util.DateTimeUtils;
 import duke.util.FileUtils;
 
 /**
- * The is the TaskList class that that
- * contains a list of task.
+ * The is the TaskList class that contains a list of tasks
+ * and the operations of the tasks.
  */
 public class TaskList {
     private static final String DIR_NAME = "data";
@@ -115,32 +115,15 @@ public class TaskList {
                 String name = contents[2];
                 switch (taskID) {
                 case Todo.ID:
-                    Todo todo = new Todo(name, isDone.equals("0"));
-                    addTask(todo);
+                    addTodoTask(isDone, name);
                     break;
                 case Deadline.ID:
                     String byTime = contents[3];
-                    Deadline deadline = new Deadline(
-                            name,
-                            DateTimeUtils.parseDateTime(byTime),
-                            isDone.equals("0")
-                    );
-                    addTask(deadline);
+                    addDeadlineTask(isDone, name, byTime);
                     break;
                 case Event.ID:
                     String atTime = contents[3];
-                    String[] dateTimes = atTime.split(" ", 2);
-                    String[] times = dateTimes[1].split("-", 2);
-                    String atDate = dateTimes[0];
-                    String startTime = times[0];
-                    String endTime = times[1];
-                    EventDateTime eventDateTime = new EventDateTime(
-                            DateTimeUtils.parseDate(atDate),
-                            DateTimeUtils.parseTime(startTime),
-                            DateTimeUtils.parseTime(endTime)
-                    );
-                    Event event = new Event(name, eventDateTime, isDone.equals("0"));
-                    addTask(event);
+                    addEventTask(isDone, name, atTime);
                     break;
                 default:
                     break;
@@ -149,6 +132,35 @@ public class TaskList {
         } catch (Exception e) {
             throw new DukeIoException("☹ OOPS!!! Load tasks from file error.");
         }
+    }
+
+    private void addTodoTask(String isDone, String name) {
+        Todo todo = new Todo(name, isDone.equals("0"));
+        addTask(todo);
+    }
+
+    private void addDeadlineTask(String isDone, String name, String byTime) {
+        Deadline deadline = new Deadline(
+                name,
+                DateTimeUtils.parseDateTime(byTime),
+                isDone.equals("0")
+        );
+        addTask(deadline);
+    }
+
+    private void addEventTask(String isDone, String name, String atTime) {
+        String[] dateTimes = atTime.split(" ", 2);
+        String[] times = dateTimes[1].split("-", 2);
+        String atDate = dateTimes[0];
+        String startTime = times[0];
+        String endTime = times[1];
+        EventDateTime eventDateTime = new EventDateTime(
+                DateTimeUtils.parseDate(atDate),
+                DateTimeUtils.parseTime(startTime),
+                DateTimeUtils.parseTime(endTime)
+        );
+        Event event = new Event(name, eventDateTime, isDone.equals("0"));
+        addTask(event);
     }
 
     /**
